pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = "AI-powered investor assistant web application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "ipykernel>=7.0.1",
    "jupyter>=1.1.1",
    "langchain-community==0.3.25",
    "langchain-core==0.3.79",
    "langchain-openai==0.3.18",
    "langchain-tavily==0.2.12",
    "langchain[openai]==0.3.25",
    "langgraph==0.4.5",
    "langgraph-prebuilt==0.2.2",
    "opik==1.8.74",
    "python-dotenv==1.1.0",
]
